Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11018 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: e958c42b0db4158c3ade6de67f0503d4b5ef4463

+++ core/relational.base/src/de/uniol/inf/is/odysseus/relational/base/predicate/TypeSafeRelationalPredicate.java
@@ -34,7 +34,7 @@ public class TypeSafeRelationalPredicate extends RelationalPredicate{
 			if(this.expression.getAllAttributes().get(i).getDatatype().getURI().equalsIgnoreCase("String") &&
 					vars.get(i).getAcceptedTypes().length == 1 ){ 
 				if(vars.get(i).getAcceptedTypes()[0].isNumeric()){
-					values[i]=(Double)Double.parseDouble((String)input.getAttribute(this.attributePositions[i]));
+					values[i]=Double.parseDouble((String)input.getAttribute(this.attributePositions[i]));
 				}
 				else{
 					values[i] = input.getAttribute(this.attributePositions[i]);
